# api/openapi/helm.openapi.yaml
openapi: "3.1.0"
info:
  title: HELM Kernel API
  version: 0.5.0
  description: |
    Deterministic execution kernel for AI tool calls.
    Drop-in OpenAI proxy + cryptographic receipts + offline-verifiable evidence packs.
  license:
    name: Apache-2.0
    url: https://github.com/Mindburn-Labs/helm-ai-kernel/blob/main/LICENSE
  contact:
    name: Mindburn Labs
    url: https://mindburn.org
servers:
  - url: http://localhost:8080
    description: Local development (Docker Compose)
  - url: http://localhost:9090
    description: Standalone proxy mode
...
paths:
...
  /api/v1/kernel/approve:
    post:
      operationId: approveIntent
      tags: [approval]
      summary: Submit an approval for a pending intent
      description: |
        RFC-005 approval ceremony endpoint.
        Requires intent_hash, Ed25519 signature, and public key.
        Returns an approval receipt with causal chain linkage.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalRequest'
            example:
              intent_hash: "sha256:abc123..."
              signature_b64: "base64-encoded-ed25519-signature"
              public_key_b64: "base64-encoded-public-key"
      responses:
        '200':
          description: Approval receipt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Receipt'
              example:
                receipt_id: "rec_a1b2c3"
                status: "APPROVED"
                reason_code: "ALLOW"
                lamport_clock: 42
                output_hash: "sha256:7f83b1..."
        '400':
          $ref: '#/components/responses/HelmError'
        '403':
          $ref: '#/components/responses/HelmError'
...
  /api/v1/receipts:
    get:
      operationId: listReceipts
      tags: [console]
      summary: List durable receipts
      security:
        - AdminBearerAuth: []
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 100, maximum: 1000 }
        - name: since
          in: query
          schema: { type: string }
        - name: agent
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Receipt list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptListResponse'
        '400':
          $ref: '#/components/responses/HelmError'
        '401':
          $ref: '#/components/responses/HelmError'
        '503':
          $ref: '#/components/responses/HelmError'
...
  /api/v1/console/bootstrap:
    get:
      operationId: getConsoleBootstrap
      tags: [console]
      summary: Read HELM AI Kernel Console bootstrap state
      security:
        - AdminBearerAuth: []
      description: |
        Consolidated read-only platform state used by the HELM AI Kernel Console shell.
        Values are derived from kernel services and local configuration.
      responses:
        '200':
          description: Console bootstrap state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsoleBootstrap'
        '401':
          $ref: '#/components/responses/HelmError'
  /api/v1/console/surfaces:
    get:
      operationId: listConsoleSurfaces
      tags: [console]
      summary: List Console surfaces and their backing source contracts
      security:
        - AdminBearerAuth: []
      responses:
        '200':
          description: Console surface catalog
          content:
            application/json:
              schema:
                type: object
                properties:
                  surfaces:
                    type: array
                    items:
                      $ref: '#/components/schemas/ConsoleSurfaceRef'
        '401':
          $ref: '#/components/responses/HelmError'
...
  /mcp:
    get:
      operationId: getMCPTransport
      tags: [mcp]
      summary: Open the MCP transport endpoint
      description: |
        Returns JSON discovery metadata by default.
        When `Accept: text/event-stream` is used, responds with an SSE primer event.
      responses:
        '200':
          description: MCP discovery metadata or SSE primer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MCPRemoteDiscovery'
            text/event-stream:
              schema:
                type: string
    post:
      operationId: postMCPJSONRPC
      tags: [mcp]
      summary: Send a JSON-RPC request to the MCP transport endpoint
      description: |
        Supports `initialize`, `notifications/initialized`, `ping`, `tools/list`, and `tools/call`.
        Protocol negotiation happens on `initialize`; later requests SHOULD include `MCP-Protocol-Version`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MCPJSONRPCRequest'
      responses:
        '200':
          description: JSON-RPC response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MCPJSONRPCResponse'
        '400':
          description: Invalid JSON-RPC body or unsupported protocol version
          content:
            text/plain:
              schema:
                type: string
...
  /api/v1/replay/verify:
    post:
      operationId: replayVerify
      tags: [evidence]
      summary: Replay and verify execution from tapes
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                bundle:
                  type: string
                  format: binary
      responses:
        '200':
          description: Replay verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationResult'
        '400':
          $ref: '#/components/responses/HelmError'

  # ── Conformance ─────────────────────────────────────
...
components:
  securitySchemes:
    AdminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: HELM_ADMIN_API_KEY
      description: |
        Self-hosted OSS admin bearer key. Set HELM_ADMIN_API_KEY on the server
        and send Authorization: Bearer <key> for protected runtime APIs.

  # ── Schemas ─────────────────────────────────────────
  schemas:
    HelmError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message, type, code, reason_code]
          properties:
            message:
              type: string
              description: Human-readable error message
            type:
              type: string
              enum: [invalid_request, authentication_error, permission_denied, not_found, internal_error]
            code:
              type: string
              description: Machine-readable error code
            reason_code:
              type: string
              description: HELM-specific reason code
              enum:
                - DENY_TOOL_NOT_FOUND
                - DENY_SCHEMA_MISMATCH
                - DENY_OUTPUT_DRIFT
                - DENY_BUDGET_EXCEEDED
                - DENY_APPROVAL_REQUIRED
                - DENY_APPROVAL_TIMEOUT
                - DENY_SANDBOX_TRAP
                - DENY_GAS_EXHAUSTION
                - DENY_TIME_LIMIT
                - DENY_MEMORY_LIMIT
                - DENY_POLICY_VIOLATION
                - DENY_TRUST_KEY_REVOKED
                - DENY_IDEMPOTENCY_DUPLICATE
                - ERROR_INTERNAL
            details:
              type: object
              description: Additional error context
              additionalProperties: true
...
    ApprovalRequest:
      type: object
      required: [intent_hash, signature_b64, public_key_b64]
      properties:
        intent_hash:
          type: string
          description: SHA-256 hash of the intent to approve
        signature_b64:
          type: string
          description: Base64-encoded Ed25519 signature
        public_key_b64:
          type: string
          description: Base64-encoded Ed25519 public key
        challenge_response:
          type: string
          description: Optional challenge/response for ceremony
...
    Receipt:
      type: object
      properties:
        receipt_id: { type: string }
        decision_id: { type: string }
        effect_id: { type: string }
        status: { type: string }
        reason_code: { type: string }
        output_hash: { type: string }
        blob_hash: { type: string }
        prev_hash: { type: string }
        lamport_clock: { type: integer }
        signature: { type: string }
        timestamp: { type: string, format: date-time }
        principal: { type: string }
        executor_id: { type: string }
        args_hash: { type: string }
        metadata:
          type: object
          additionalProperties: true
    ReceiptListResponse:
      type: object
      properties:
        receipts:
          type: array
          items:
            $ref: '#/components/schemas/Receipt'
        count: { type: integer }
        next_cursor: { type: string }
        has_more: { type: boolean }
...
    ConsoleBootstrap:
      type: object
      required: [version, workspace, health, counts, receipts, conformance, mcp]
      properties:
        version:
          $ref: '#/components/schemas/VersionInfo'
        workspace:
          type: object
          required: [organization, project, environment, mode]
          properties:
            organization: { type: string }
            project: { type: string }
            environment: { type: string }
            mode: { type: string }
        health:
          type: object
          required: [kernel, policy, store, conformance]
          properties:
            kernel: { type: string }
            policy: { type: string }
            store: { type: string }
            conformance: { type: string }
        counts:
          type: object
          required: [receipts, pending_approvals, open_incidents, mcp_tools]
          properties:
            receipts: { type: integer }
            pending_approvals: { type: integer }
            open_incidents: { type: integer }
            mcp_tools: { type: integer }
        receipts:
          type: array
          items:
            $ref: '#/components/schemas/Receipt'
        conformance:
          type: object
          required: [level, status]
          properties:
            level: { type: string }
            status: { type: string }
            report_id: { type: string }
        mcp:
          type: object
          required: [authorization, scopes]
          properties:
            authorization: { type: string }
            scopes:
              type: array
              items: { type: string }
    ConsoleSurfaceRef:
      type: object
      required: [id, source]
      properties:
        id: { type: string }
        source: { type: string }
...
    MCPRemoteDiscovery:
      type: object
      properties:
        server_name: { type: string }
        version: { type: string }
        capabilities_url: { type: string }
        execute_url: { type: string }
        mcp_endpoint: { type: string }
        supported_protocol_versions:
          type: array
          items: { type: string }
        auth_mode: { type: string }
        governance: { type: string }
    MCPJSONRPCRequest:
      type: object
      required: [jsonrpc, method]
      properties:
        jsonrpc:
          type: string
          enum: ["2.0"]
        id:
          description: Request identifier. Omit for notifications.
          oneOf:
            - type: string
            - type: integer
        method: { type: string }
        params:
          type: object
          additionalProperties: true
    MCPJSONRPCError:
      type: object
      properties:
        code: { type: integer }
        message: { type: string }
    MCPJSONRPCResponse:
      type: object
      properties:
        jsonrpc:
          type: string
          enum: ["2.0"]
        id:
          oneOf:
            - type: string
            - type: integer
        result:
          type: object
          additionalProperties: true
        error:
          $ref: '#/components/schemas/MCPJSONRPCError'
...
    VerificationResult:
      type: object
      properties:
        verdict:
          type: string
          enum: [PASS, FAIL]
        checks:
          type: object
          properties:
            signatures: { type: string, enum: [PASS, FAIL] }
            causal_chain: { type: string, enum: [PASS, FAIL] }
            policy_compliance: { type: string, enum: [PASS, FAIL] }
        roots:
          type: object
          properties:
            manifest_root_hash: { type: string }
            merkle_root: { type: string }
            entry_count: { type: integer }
        errors:
          type: array
          items: { type: string }
...
    VersionInfo:
      type: object
      properties:
        version: { type: string }
        commit: { type: string }
        build_time: { type: string }
        go_version: { type: string }
...
  responses:
    HelmError:
      description: HELM error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HelmError'
          example:
            error:
              message: "Tool not found: unknown_tool"
              type: "invalid_request"
              code: "tool_not_found"
              reason_code: "DENY_TOOL_NOT_FOUND"